pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "shtops"
version = "0.1.0"
description = "SHTops operational independence context layer"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "requests>=2.31.0",
  "pyyaml>=6.0",
  "python-dateutil>=2.8.2",
  "flask>=3.0.0",
]

[project.scripts]
shtops = "shtops.cli:main"

[tool.setuptools]
package-dir = {"" = "."}

[tool.setuptools.packages.find]
where = ["."]
include = ["shtops*"]
